feature[FE-PCOMP-1] Inicio da seleção personalizada de filtro de estado para a pagina de comparação
