+++ src/components/Maps/Maps.module.css
@@ -1,17 +1,4 @@
 
-/* BrazilMap */
-.mapCurrentState{
-    margin-top: 0.875rem;
-    color: var(--black-500);
-    font: var(--title-regular);
-    text-align: center;
-  }
-  
-  .mapDescription {
-    color: var(--black-500);
-    font: var(--text-medium);
-    text-align: center;
-  }
   
 /* WorldMap */
 rect{
