+++ src/pages/comparisonStats/ComparisonStats.jsx
@@ -16,6 +16,22 @@ import Dropdown from '../../components/Dropdown/Dropdown'
 import { faCircleInfo } from "@fortawesome/free-solid-svg-icons"
 
 const ComparisonStats = () => {
+
+    const [region, setRegion] = useState('');
+    const [state, setState] = useState('');
+
+    // Opções de descrição para o mapa do Brasil
+    const getDescriptionText = () => {
+        if (state) { //Selecionou um estado
+        return "Para desfazer a seleção de estado atual, clique no mapa abaixo.";
+        } else if (region) { //Selecionou uma região
+        return "Escolha um dos estados para analisar seus dados.";
+        } else { //Não selecionou nada
+        return "Para selecionar um estado, escolha uma das regiões do mapa abaixo.";
+        }
+    };
+
+
     // TESTE ATUALIZAÇÃO DO ESTADO YEAR
     const dadosTeste = {
       exportacao: [
@@ -141,7 +157,6 @@ const ComparisonStats = () => {
         },
       ]
     };
-
     
     // Variáveis para os inputs
     const [product , setProduct] = useState('')
@@ -250,7 +265,18 @@ const ComparisonStats = () => {
 
             <section id={styles.primaryInfos}>
                 <div className={styles.navMap}>
-                    <BrazilMap/>
+                    
+                    {/* Legenda do mapa do brasil */}
+                    <p className={styles.mapDescription}>{getDescriptionText()}</p>
+
+                    {/* Região/Estado selecionado */}
+                    {state ? (
+                    <h2 className={styles.mapCurrentState}>{state}</h2>
+                    ) : region ? (
+                    <h2 className={styles.mapCurrentState}>Região {region}</h2>
+                    ) : null}
+
+                    <BrazilMap onRegionChange={({ regiao, estado }) => { setRegion(regiao ?? undefined); setState(estado ?? undefined);}} />
                 </div>
 
                 <section className={`${styles.infoGridVertical} infoGridVertical`}>
